mgr/cephadm: mark host is offline in osd rm status
If a host we are trying to remove the OSD on is
offline, removal of the OSD will never totally
complete as the OSD container itself cannot
be cleaned up by cephadm. This means the status
shown in "ceph orch osd rm status" will just
sit at "started", "draining", etc. permanently.
Instead we should report in the status that
the host the OSD is on is offline so users have
a better idea why the removal is not finishing.

Fixes: https://tracker.ceph.com/issues/61663

Signed-off-by: Adam King <adking@redhat.com>