mds: fix deadlock between unlinking and reintegrating
The unlink request could trigger an inode to do the linkmerge, if
so when the unlink request could set the dentry's state to
STATE_UNLINKING when tries first time, and then for the second time
to try the unlink request it detected the corresponding dentry was
being linkmerged, so the unlink will goto sleep and waiting for
the linkmerging to finish.

While for the linkmerge it will be a internal rename request, which
will goto sleep to because the dentry't state is STATE_UNLINKING.

This fix will add one intemediate state and always let linkmerge
to be finished first.

Signed-off-by: Xiubo Li <xiubli@redhat.com>