ceph-volume: fix raw activate by osd-id/osd-uuid
When activating a raw osd with `--osd-id` and `--osd-uuid` instead of
passing the device, it ends up with a call to List.generate() passing
`[None]` (which doesn't really make sense).
When no devices are passed tp List.generate(), it needs to enter the
first condition in order to build a list of all devices. Otherwise,
that function starts iterating on an empty list which result in a
failure (see corresponding tracker).

Fixes: https://tracker.ceph.com/issues/61855

Signed-off-by: Guillaume Abrioux <gabrioux@ibm.com>