pybind/mgr/cephadm: move haproxy client addrs to func
Move the list comprehension that gets the list of IP addresses for the
haproxy_hosts list to a separate function. This is in preparation for
making the logic more complex in a later commit.

Signed-off-by: John Mulligan <jmulligan@redhat.com>
(cherry picked from commit b1d2bb6ec53c2aa3574db2f5d3dc3878d2969a58)