cephadm: rename the requirements file for the zipapp
Previously, we had code in the build.py script that looked for
`requirements.txt` even though there was no requirements file in
cephadm. In preliminary testing, I found that using pip to get
dependency packages for a zipapp we need to behave a bit special - the
zipapp should not contain binary, architecture specific, loadable
modules. Because the needs of a portable zipapp are special we give the
new special name to our requirements file to help act as a reminder of
this special-ness.

Signed-off-by: John Mulligan <jmulligan@redhat.com>