common/TrackedOp: fix race updating description with proper lock
It is not safe to "cache" a member which may be changed by a racing thread.

This reworks the locking so we can do a light-weight check if the description
is already generated wihtout acquiring the heavier TrackedOp::lock. If it's not
available yet or it needs regenerated, then get the proper locks to generate
it.

Fixes: e45f5c2c33508bec3b177e51fd6cc5fde1f0e674
Signed-off-by: Patrick Donnelly <pdonnell@redhat.com>
(cherry picked from commit 36eaa495eb37485796ce57c8c4b1cca603efc193)