REMOVE ME: set nvmeof image to testing image
The current nvmeof image we have only works with
admin keyring. This is a placeholder commit that
points to an image with changes that should allow
the nvmeof daemon to use the keyring we generate
for it. This should be removed once a proper nvmeof
image is made with the necessary changes from
https://github.com/ceph/ceph-nvmeof/pull/166