qa: pass arg as list to fix test case failure
Pacific branch's run_cluster_cmd() code can't split string into a list, and
because backporting is too tricky as a lot of qa code doesn't reach pacific, it
would be tedious to backport the PRs that enabled passing string as arg to
run_cluster_cmd() in ceph_manager.py. For complete details please checkout
note #13 in the tracker attached below.

Fixes: https://tracker.ceph.com/issues/61732
Signed-off-by: Dhairya Parmar <dparmar@redhat.com>