qa/cephadm: mgr-nfs-upgrade, match any migration > 2
I believe this check was originally added because
the 2->3 migration migrated some nfs related bits. Since
then we've had to update the migration this checks
for every time we bump the max migration. This change
is intended to instead just have it check for a
miration > 2 so we don't have to keep updating it.

It seemed to work in that it only gave a 0 return
code on integers > 2

[ceph: root@vm-00 /]# echo "error" | awk '$1 > 2' | grep -E '^[1-9]*$'; echo $?
1
[ceph: root@vm-00 /]# echo "error 6" | awk '$1 > 2' | grep -E '^[1-9]*$'; echo $?
1
[ceph: root@vm-00 /]# echo "6 error" | awk '$1 > 2' | grep -E '^[1-9]*$'; echo $?
1
[ceph: root@vm-00 /]# echo "" | awk '$1 > 2' | grep -E '^[1-9]*$'; echo $?
1
[ceph: root@vm-00 /]# echo "1" | awk '$1 > 2' | grep -E '^[1-9]*$'; echo $?
1
[ceph: root@vm-00 /]# echo "-7" | awk '$1 > 2' | grep -E '^[1-9]*$'; echo $?
1
[ceph: root@vm-00 /]# echo "3" | awk '$1 > 2' | grep -E '^[1-9]*$'; echo $?
3
0
[ceph: root@vm-00 /]# echo "15" | awk '$1 > 2' | grep -E '^[1-9]*$'; echo $?
15
0

and we expect the return value of the config get to be just an integer
like "6" unless it errors for some reason

Signed-off-by: Adam King <adking@redhat.com>