osd: check if adjacent clones are unreadable when rollback is called
In rollback, if the head object is not manfest and the rollback_to is manifest,
the head object becomes the manifest object. At this point,
we need to check adjacent clonces to the head object to calculate
correct reference count for deduped chunks because the head object is now manifest.
Plus, in the meantime of the waiting for recovery, subsequent requests should be blocked
to ensure transaction order.

To fix the issue, this commit is to wait until the adjacent
clones are readable and delay future incoming ops.

fixes: https://tracker.ceph.com/issues/62167

Signed-off-by: Myoungwon Oh <myoungwon.oh@samsung.com>