crimson/osd: simplify, rename SnapTrimMutex, remove phase wrapper
There wasn't much utility in wrapping the shared_mutex with an
OrderedConcurrentPhase.  It didn't really work as a way of
registering the op as waiting on the mutex since the op enters
the stage prior to blocking on the mutex.

Moreover, we want to disallow scrub from operating concurrently
with trim for now -- see the comment for an explanation.

Thus, let's just simplify it to a mutex.  We can add the BlockerT
machinery later.

Signed-off-by: Samuel Just <sjust@redhat.com>