osd: Apply randomly selected scheduler type across all OSD shards
Originally, the choice of 'debug_random' for osd_op_queue resulted in the
selection of a random scheduler type for each OSD shard. A more realistic
scenario for testing would be the selection of the random scheduler type
applied globally for all shards of an OSD. In other words, all OSD shards
would employ the same scheduler type. For e.g., this scenario would be
possible during upgrades when the scheduler type has changed between
releases.

The following changes are made as part of the commit:
 1. Determine the scheduler type before initializing the OSD shards in
    OSD class constructor.
 2. Maintain the scheduler type string (osd_op_queue) within the OSD class
 3. Pass the scheduler type string to the OSDShard's make_scheduler()
    method for each shard. This ensures all shards of the OSD are
    initialized with the same schedueler type.

Fixes: https://tracker.ceph.com/issues/62171
Signed-off-by: Sridhar Seshasayee <sseshasa@redhat.com>