osd/OSD: Use OSD's osd_op_queue instead of querying config subsystem
With the osd_op_queue string a member of the OSD class, the same may be
used where applicable to determine the scheduler type. This results in the
appropriate setting of different config options based on the scheduler
type in case the global config option is set to 'debug_random' (for e.g.
in CI tests).

Note that the use of the OSD's osd_op_queue applies only for the methods
of the OSD class. In other places like for e.g., PG specific code, the
global config option for osd_op_queue ('debug_random') would continue to
be used.

https://tracker.ceph.com/issues/62171
Signed-off-by: Sridhar Seshasayee <sseshasa@redhat.com>