rgw/ops-log: explicitly specify object name in the log entry
Pseudo-directories can be used when naming an object (key): e.g.,
"my/weird/object.txt". This would lead to cases where the name of the object cannot be derived deterministically. For example, if an object name uses pseudo-directories and starts with "<bucket_name>/" and virtual-host style is used when accessing object, we cannot figure out the name of the object. Note that in DNS (virtual host) style bucket naming, since bucket name is specified as a part of the host name, URI doesn't contain any reference to the bucket.

boto_client = boto3.client(..., config=Config(s3={'addressing_style': 'virtual'}))

boto_client.put_object(
        Body=b"this is the data",
        Key="my-bucket/my-object",
        Bucket="my-bucket"
)

The corresponding log entry is

{   ...,
    "bucket":"my-bucket", ...,
    "uri":"PUT /my-bucket/my-object HTTP/1.1",
    ...
}

We can falsely conclude that the name of the object is "my-object".

By having the name of the object listed in the log-entry, we address
this ambiguity.

Signed-off-by: Oguzhan Ozmen <oozmen@bloomberg.net>