mds: do not simplify fragset
Problem:
Frags in simplified fragset aren't found as is on the replica during scrub.

fragset::simplify() computes the bit representation of the least common
ancestor of the frags.  When this representation is forwarded to the replicas,
the ScrubStack::handle_scrub() method in the OP_QUEUEDIR case simply searches
this simplified representation in the list of (unsimplified) frags ...
which it fails to find.
Hence we get to see the "no frag 10*" type of logs.

Solution:
Do not simplify fragset when forwarding fragset to replica for scrub.

Fixes: https://tracker.ceph.com/issues/62658
Signed-off-by: Milind Changire <mchangir@redhat.com>
(cherry picked from commit f3c0424bb583423db80009c29f73e484aad840b4)