cephadm: convert ceph exporter type to a ContainerDaemonForm
CephExporter was being (partially) over-shadowed by the Ceph class as
the Ceph class listed 'ceph-exporter' as one of the daemon types it
handled. This change updates CephExporter to a ContainerDaemonForm while
simultaneously breaking the link between Ceph and 'ceph-exporter',
allowing CephExporter to handle all the duty of managing ceph-exporter,
continuing the process of having clearer logical responsibilities and
class hierarchy in cephadm.

Signed-off-by: John Mulligan <jmulligan@redhat.com>