qa/cephadm: use abs path for cephadm in test_repos test
We've been hitting issues related to jinja2 in the
cephadm zipapp if it is called using a relative path
(works fine when using abs path or when placing it
somewhere like /usr/sbin and relying on the PATH env
variable to find it). When using a relative path,
jinja2 seems to be trying to find templates based
on the relative location of the zipapp rather than
within the zipapp itslef. For example, we were hitting

Traceback (most recent call last):
  File "/tmp/tmp7zwwodz0.cephadm.build/jinja2/loaders.py", line 357, in get_source
  File "<frozen zipimport>", line 179, in get_data
'/home/ubuntu/cephtest/mnt.0/client.0/tmp/tmp.test_repos.sh.YJZ6EN/tmp.cephadm.lYKw39/cephadmlib/templates/cephadm.logrotate.config.j2'

when the zipapp was located within
/home/ubuntu/cephtest/mnt.0/client.0/tmp/tmp.test_repos.sh.YJZ6EN/
and it was being called with a relative path

Signed-off-by: Adam King <adking@redhat.com>