osd: do not promote evicted data when reading a manifest chunked object
When handling a read request for a manifest chunked object,
the current implementation issues sub-read ops to read underlying chunks
only if the requested range is within the chunk ranges the manifest object has.
If the requested range extends beyond any of the chunk ranges,
promote_object() is called to read all missing chunks and write those chunks
to the manifest object, followed by processing the read operation.

But, this behavior unnecessarily promotes the chunks that
are mared as missing state.

To address this, this commit modifies behavior to issue sub-read ops
if the requested range includes at least a chunk range
without promotion.

Signed-off-by: Myoungwon Oh <myoungwon.oh@samsung.com>