rbd: data corruption when mirroring clones
Parent data is not synced to the non-primary with snapshot based rbd-mirroring
when the primary clone image is written to. The rbd-mirror daemon uses
deep_copy with subsets of snapshots. If a copyup operation on the clone image
modifies the already synced older snapshot with the parent data, a snapshot
diff will only return the new data that was written to the clone.
The fix checks that the parent image snap from which the clone was created
has been synced before creating the non-primary clone image and triggers
a copyup as part of the object copy.

Fixes: https://tracker.ceph.com/issues/61891

Signed-off-by: N Balachandran <nibalach@redhat.com>