mon/OSDMonitor: always create pgs in the epoch in which the pool was created
The logic here didn't actually work.  Marking the pg history with an
epoch other than the epoch at which the pool was created in general
causes the PG's pg_history_t::same_interval_since value to be wrong
for the first interval.  This can cause an IO submitted by a client
based on the epoch in which the pool was created to be rejected by
the OSD without an interval change actually occuring and therefore
without the client resending the op.

In order to make this limit actually function, once we actually process
a pending pg, we'd have to go back to the OSDMap at which the pool was
created and work forward to get the correct interval bound.  That seems
even more expensive, so instead this patch simply removes the limit.

Fixes: https://tracker.ceph.com/issues/64546
Signed-off-by: Samuel Just <sjust@redhat.com>