ceph.spec.in: use gcc-toolset 13
since gts 13 is out, and the newer the toolchain the merrier. and
because clang prefers using a newer gts when its gcc is around. so
let's bump up the gcc-toolset from 11 to 13.

since annobin plugin package was renamed to
gcc-toolset-13-gcc-plugin-annobin, let's update its name accordingly.

and use -runtime subpackage instead of -build, as
macros.gcc-toolset-13-enable is now located in -runtime subpackage
since devtoolset12

Signed-off-by: Kefu Chai <tchaikov@gmail.com>